+++ app/Repositories/FeedbackRepository.php
@@ -4,6 +4,8 @@ namespace App\Repositories;
 
 use App\Contracts\IFeedbackRepository;
 use App\DTO\FeedbackDTO;
+use App\Exceptions\BusinessException;
+use App\Exceptions\ModelNotFoundException;
 use App\Models\Feedback;
 use Illuminate\Http\JsonResponse;
 
@@ -25,7 +27,11 @@ class FeedbackRepository implements IFeedbackRepository
     public function getFeedbackById(int $feedbackId): ?Feedback
     {
 
-        return Feedback::query()->find($feedbackId);
+        $feedback = Feedback::find($feedbackId);
+        if (!$feedback) {
+            throw new ModelNotFoundException(__('feedbacks.feedback_not_found'));
+        }
+        return $feedback;
     }
 
     /**
@@ -34,43 +40,56 @@ class FeedbackRepository implements IFeedbackRepository
      */
     public function createFeedback(FeedbackDTO $feedbackDTO): Feedback
     {
-        $feedback = new Feedback();
-        $feedback->user_id = $feedbackDTO->getUserId();
-        $feedback->hotel_id = $feedbackDTO->getHotelId();
-        $feedback->description = $feedbackDTO->getDescription();
-        $feedback->rating = $feedbackDTO->getRating();
-
-        $feedback->save();
-        return $feedback;
+        try {
+            $feedback = new Feedback();
+            $feedback->user_id = $feedbackDTO->getUserId();
+            $feedback->hotel_id = $feedbackDTO->getHotelId();
+            $feedback->description = $feedbackDTO->getDescription();
+            $feedback->rating = $feedbackDTO->getRating();
+            $feedback->save();
+            return $feedback;
+        }  catch (\Exception $e) {
+            throw new BusinessException(__('feedbacks.failed_to_create_feedback'));
+        }
     }
 
     /**
      * @param FeedbackDTO $feedbackDTO
      * @param int $feedbackId
      * @return Feedback
+     * @throws ModelNotFoundException
      */
     public function updateFeedback(FeedbackDTO $feedbackDTO, int $feedbackId): Feedback
     {
         $feedback = Feedback::find($feedbackId);
+        if (!$feedback) {
+            throw new ModelNotFoundException(__('feedbacks.feedback_not_found'));
+        }
+        try {
+            $feedback->user_id = $feedbackDTO->getUserId();
+            $feedback->hotel_id = $feedbackDTO->getHotelId();
+            $feedback->description = $feedbackDTO->getDescription();
+            $feedback->rating = $feedbackDTO->getRating();
+            $feedback->update();
+            return $feedback;
 
-        $feedback->user_id = $feedbackDTO->getUserId();
-        $feedback->hotel_id = $feedbackDTO->getHotelId();
-        $feedback->description = $feedbackDTO->getDescription();
-        $feedback->rating = $feedbackDTO->getRating();
-
-        $feedback->update();
-
-        return $feedback;
+        }  catch (\Exception $e) {
+            throw new BusinessException(__('feedbacks.failed_to_create_feedback'));
+        }
     }
 
     /**
      * @param int $feedbackId
      * @return void
+     * @throws ModelNotFoundException
      */
     public function destroyFeedback(int $feedbackId)
     {
 
         $feedback = Feedback::find($feedbackId);
+        if (!$feedback) {
+            throw new ModelNotFoundException(__('feedbacks.feedback_not_found'));
+        }
         $feedback->delete();
     }
